+++ src/main/java/org/spin/grpc/service/form/import_file_loader/ImportFileLoaderConvertUtil.java
@@ -13,7 +13,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.            *
  ************************************************************************************/
 
-package org.spin.form.import_file_loader;
+package org.spin.grpc.service.form.import_file_loader;
 
 import org.adempiere.core.domains.models.I_AD_Column;
 import org.adempiere.core.domains.models.I_AD_Table;
