+++ src/main/java/org/spin/grpc/service/form/import_file_loader/ImportFileLoaderServiceLogic.java
@@ -12,7 +12,7 @@
  * You should have received a copy of the GNU General Public License                *
  * along with this program. If not, see <https://www.gnu.org/licenses/>.            *
  ************************************************************************************/
-package org.spin.form.import_file_loader;
+package org.spin.grpc.service.form.import_file_loader;
 import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
